Moved (test) functions into manipulations
